Add files via upload
Incident Resolution Breakdown: A pie chart categorizing the resolution of incidents, indicating that 75.24% of cases remain open or active, while the rest are divided among other categories such as adult arrests and unfounded incidents.

Crime Statistics by District Police and Month: A detailed table that presents the total number of crimes and the percentage of solved cases  over the months of the year. For example, in January, Bayview reported 2,197 crimes with a 22% solve rate, while Central reported 3,904 crimes with a 21% solve rate and so on.

Annual Crime Overview: A bar chart comparing crime data across two years. The data shows a significant decrease of 55.11% in total crimes from 2018 (154.58K) to 2019 (69.38K).

Crimes by Time of Day: A horizontal bar chart illustrating crime distribution throughout different times of the day, showing that the highest number of crimes occurred at noon (61.64K), while the lowest number was recorded at midnight (33.41K).

Monthly Crime Trend: A line graph depicting the trend of crime across months, indicating that crime peaked in May (25.1K) and significantly decreased in the following months, with the lowest point in October (12.1K).